Split utls library
