+++ transport/internet/tls/grpc.go
@@ -7,7 +7,7 @@ import (
 	"net/url"
 	"strconv"
 
-	utls "github.com/sagernet/utls"
+	utls "github.com/refraction-networking/utls"
 	"google.golang.org/grpc/credentials"
 )
